Avance polinomios

+++ src/main/java/com/proyecto/OperacionesPolinomios.java
@@ -8,9 +8,9 @@ public class OperacionesPolinomios {
         int n= p1.getGrado();
         int m=p2.getGrado(),i=0;
         int grado = (n>=m) ? n : m;
-        int[] coef= new int[grado+1];
+        double [] coef= new double[grado+1];
         while(m>=0 || n >= 0){
-            int p11, p22;
+            double p11, p22;
             if(n<0)
                 p11=0;
             else
@@ -32,9 +32,9 @@ public class OperacionesPolinomios {
         int n= p1.getGrado();
         int m=p2.getGrado(),i=0;
         int grado = (n>=m) ? n : m;
-        int[] coef= new int[grado+1];
+        double[] coef= new double[grado+1];
         while(m>=0 || n >= 0){
-            int p11, p22;
+            double p11, p22;
             if(n<0)
                 p11=0;
             else
@@ -55,7 +55,7 @@ public class OperacionesPolinomios {
             throw new NullPointerException("Debes introducir dos polinimios para multoplicarr");
         int n= p1.getGrado();
         int m=p2.getGrado();
-        int[] coef= new int[n+m+1];
+        double[] coef= new double[n+m+1];
         for(int i=0;i<n+1;i++){
             for (int j=0;j<m+1;j++){
                 coef[i+j]=coef[i+j]+(p1.getCoeficientes()[i]*p2.getCoeficientes()[j]);
